processamento-de-imagens/trabalho/api/requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
fastapi==0.78.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
numpy==1.21.6
Pillow==9.1.1
pydantic==1.9.1
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
